feat(pawz): added utoipa bootstrap
