+++ crates/pawz/Cargo.toml
@@ -7,7 +7,11 @@ authors.workspace = true
 
 [dependencies]
 axum = { workspace = true }
+reqwest = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["net"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
+utoipa = { workspace = true, features = ["axum_extras", "debug"] }
+utoipa-axum = { workspace = true, features = ["debug"] }
+utoipa-swagger-ui = { workspace = true, features = ["axum", "reqwest"] }
