+++ crates/pawz/src/lib.rs
@@ -1,11 +1,13 @@
 #![allow(dead_code)]
-use axum::Router;
 use error::Error;
 use std::fmt::{Debug, Display};
 use std::net::Ipv4Addr;
 use std::ops::RangeInclusive;
 use tokio::net::TcpListener;
 use tracing::{debug, error, info, instrument};
+use utoipa::OpenApi;
+use utoipa_axum::router::OpenApiRouter;
+use utoipa_swagger_ui::SwaggerUi;
 
 pub mod config;
 pub mod error;
@@ -35,6 +37,10 @@ impl Display for Environment {
         }
     }
 }
+
+#[derive(OpenApi)]
+#[openapi()]
+struct ApiDoc;
 
 /// Represents the application configuration.
 ///
@@ -147,7 +153,7 @@ where
     ///
     /// A `Result` indicating success or an `Error` on failure.
     #[instrument]
-    pub async fn run(self, routes: Router) -> Result<(), Error> {
+    pub async fn run(self, routes: fn() -> OpenApiRouter) -> Result<(), Error> {
         debug!("running {}:{} in {}", &self.name, &self.version, &self.env);
         if let Some(tracer) = &self.tracer {
             tracer.setup(&self.env);
@@ -175,6 +181,13 @@ where
             }
         };
 
+        debug!("creating open api routes");
+        let (router, api) = OpenApiRouter::with_openapi(ApiDoc::openapi())
+            .merge(routes())
+            .split_for_parts();
+
+        let routes = router.merge(SwaggerUi::new("/swagger").url("/api-docs/openapi.json", api));
+
         debug!(
             "starting {} on {}",
             &self.name,
@@ -191,7 +204,8 @@ where
     }
 }
 
-// TODO: add openapi utoipa here.
+// TODO: updated openapi utoipa description, info and some texts after runtime
+// TODO: divide stuff into tracing_config, openapi_config, general_config
 // TODO: add tests to see if this will work.
 // TODO: enable prod, default is dev unlesss specified differently.?
 // TODO: add prod tracing setup.
